+++ python/aitemplate/frontend/nn/dropout.py
@@ -24,7 +24,7 @@ class Dropout(Module):
     def __init__(
         self,
         p=0,
-        dtype="float16",
+        dtype="float32",
     ):
         super().__init__()
 
@@ -40,6 +40,6 @@ class DropPath(Dropout):
 
     def __init__(
         self,
-        dtype="float16",
+        dtype="float32",
     ):
         super().__init__()
